+++ src/utils/constants.ts
@@ -1,2 +1,10 @@
 export const INCREASED_TEST_TIMEOUT = 30 * 1000;
 // default test timeout is 5 seconds, that's not enough for certain tests, so we increase it to 30 seconds
+
+export const BREAKPOINTS = {
+  xsm: 0,
+  sm: 361,
+  md: 961,
+  lg: 1401,
+  xl: 1801,
+};