+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.54.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.53.1...v1.54.0) (2024-09-13)
+
+
+### Features
+
+* **seldonprovider:** L3-3725 add new seldon provider component that provides media query context ([#307](https://github.com/PhillipsAuctionHouse/seldon/issues/307)) ([c5334ef](https://github.com/PhillipsAuctionHouse/seldon/commit/c5334ef79ee5f47a7588a484c7ea70f56de60fd2))
+
+## [1.53.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.53.0...v1.53.1) (2024-09-12)
+
+
+### Bug Fixes
+
+* **login:** L3-3694 increase mobile icon size ([#306](https://github.com/PhillipsAuctionHouse/seldon/issues/306)) ([18c8977](https://github.com/PhillipsAuctionHouse/seldon/commit/18c8977b3e27ec4a3421664cf9a325867dc943e7))
+
+# [1.53.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.52.0...v1.53.0) (2024-09-11)
+
+
+### Features
+
+* **readmore:** L3-3579 Read More and Collapsible Component ([#295](https://github.com/PhillipsAuctionHouse/seldon/issues/295)) ([ef3e7ca](https://github.com/PhillipsAuctionHouse/seldon/commit/ef3e7ca9b72b08fdeb635b1206274e6f8f0b90c6))
+
 # [1.52.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.51.1...v1.52.0) (2024-09-10)